feat: pass basic artists.e2e tests

+++ src/artists/artists.controller.ts
@@ -1,6 +1,10 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, Put, Res } from '@nestjs/common';
 import { ArtistsService } from '../artists/artists.service';
-import { Artist } from '../interfaces/artists.interface';
+import { CreateArtistDto } from '../interfaces/artists.interface';
+import { Response } from 'express';
+import { validate } from 'uuid';
+import { StatusCodes } from 'http-status-codes';
+
 
 @Controller('artist')
 export class ArtistsController {
@@ -8,8 +12,13 @@ export class ArtistsController {
   constructor(private artistsService: ArtistsService) {}
 
   @Post()
-  createArtist() {
-    return this.artistsService.createArtist();
+  createArtist(@Body() body: CreateArtistDto, @Res({ passthrough: true }) res: Response) {
+    if (!body?.name || !body?.grammy) {
+      res.status(StatusCodes.BAD_REQUEST).send('Required fields are not filled in');
+      return;
+    } else {
+      return this.artistsService.createArtist(body);
+    }
   }
 
   @Get()
@@ -18,18 +27,47 @@ export class ArtistsController {
   }
 
   @Get(':id')
-  getArtistById(@Param('id') id: string) {
-    return this.artistsService.getArtistById(id);
+  getArtistById(@Param('id') id: string, @Res({ passthrough: true }) res: Response) {
+    if (!validate(id)) {
+      res.status(StatusCodes.BAD_REQUEST).send('Artist id is invalid');
+      return;
+    } else if (!this.artistsService.getArtistById(id)) {
+      res.status(StatusCodes.NOT_FOUND).send('Artist does not exist');
+      return;
+    } else {
+      return this.artistsService.getArtistById(id);
+    }
   }
 
   @Put(':id')
-  updateArtistInfo(@Param('id') id: string, @Body() body: Artist) {
+  updateArtistInfo(@Param('id') id: string, @Body() body: CreateArtistDto, @Res({ passthrough: true }) res: Response,) {
+    if (!validate(id)) {
+      res.status(StatusCodes.BAD_REQUEST).send('Artist id is invalid');
+      return;
+    }
+    // if (!body?.name || !body?.grammy || typeof body?.name !== 'string') {
+    //   res.status(StatusCodes.BAD_REQUEST).send('Artist dto is invalid');
+    //   return;
+    // }
+    if (!this.artistsService.getArtistById(id)) {
+      res.status(StatusCodes.NOT_FOUND).send('Artist does not exist');
+      return;
+    }
     return this.artistsService.updateArtistInfo(id, body);
   }
 
   @Delete(':id')
-  deleteArtist(@Param('id') id: string) {
-    return this.artistsService.deleteArtistById(id)
+  deleteArtist(@Param('id') id: string, @Res() res: Response) {
+    if (!validate(id)) {
+      res.status(StatusCodes.BAD_REQUEST).send('Artist id is invalid');
+      return;
+    }
+    if (!this.artistsService.getArtistById(id)) {
+      res.status(StatusCodes.NOT_FOUND).send('Artist does not exist');
+      return;
+    }
+    this.artistsService.deleteArtistById(id);
+    res.status(StatusCodes.NO_CONTENT).send();
+    return;
   }
-
 }
